More menu style refactoring
