+++ CSETWebNg/src/app/layout/acet-layout-main/acet-layout-main.component.scss
@@ -33,8 +33,8 @@ $black: #000000;
 
 /* --- Blues --- */
 $primary: ( 50: #E9EDF2, 100: #D1D9E6, 150: #BCCFDA, 200: #A3B4CC, 250: #8FB0C1, 300: #79A0B5, 350: #6390A9,
-400: #83A5D4, 450: #215283, 500: #008ABB, 550: #0A5278, 600: #0A4B6E, 650: #094463, 700: #2c5c6d,
-750: #083C58, 800: #07354D, 850: #062D42, 900: #17395C, 950: #041E2C);
+400: #83A5D4, 450: #3B68AA, 500: #206184, 550: #0A5278, 600: #0A4B6E, 650: #094463, 700: #2c5c6d,
+750: #083C58, 800: #07354D, 850: #062D42, 900: #18375F, 950: #041E2C);
 
 $primary-50: map-get($primary, 50);
 $primary-100: map-get($primary, 100);
@@ -55,9 +55,9 @@ $primary-800: map-get($primary, 800);
 $primary-850: map-get($primary, 850);
 $primary-900: map-get($primary, 900); // Main Dark Primary
 
-$primary-color: $primary-450;
+$primary-color: $primary-900;
 
-$secondary-color: $primary-900;
+$secondary-color: $primary-450;
 
 $link-color-dark: #1a53ff; //$primary-350;
 $link-color-light: $primary-350;
@@ -131,7 +131,7 @@ body {
 
 .nav-link:focus,
 .navbar-menu-header:focus {
-  border:1px solid #3B68AA  !important;
+  border: 1px solid #778baa !important;
   border-radius: 4px !important;
   text-decoration: none !important;
 }
@@ -194,33 +194,6 @@ a, a:not([href]) {
   color: $black;
 }
 
-.btn-link-subtle {
-  font-weight: 400;
-  color: $primary-color;
-  background-color: transparent;
-}
-.btn-link-subtle:hover {
-  color: $primary-color;
-  text-decoration: underline dotted;
-  background-color: transparent;
-  border-color: transparent;
-}
-.btn-link-subtle:focus,
-.btn-link-subtle.focus {
-  text-decoration: underline dotted;
-  border-color: transparent;
-  box-shadow: none;
-}
-.btn-link-subtle:disabled,
-.btn-link-subtle.disabled {
-  color: #6c757c;
-  pointer-events: none;
-}
-
-.no-underline {
-  text-decoration: none !important;
-}
-
 h2,
 h3,
 h4 {
@@ -234,7 +207,7 @@ h4 {
 }
 
 label[required]::after {
-  content: " *";
+  content: ' *';
   font-weight: bold;
   color: $validation-color;
 }
@@ -252,7 +225,7 @@ label[required]::after {
   left: 0;
   right: 0;
   margin: 0 auto;
-  opacity: 0.1;
+  opacity: .1;
 }
 
 .bg-semi-trans {
@@ -272,7 +245,6 @@ label[required]::after {
 /* --- header --- */
 .cset-header {
   background-color: $primary-color;
-  background-color: #215283;
 }
 
 /* --- footer --- */
@@ -318,7 +290,7 @@ label[required]::after {
 
 .footer-section-2 {
   color: $gray-50;
-  background-color: $primary-color;
+  background-color: $secondary-color;
   line-height: 2em;
 }
 
@@ -349,7 +321,7 @@ label[required]::after {
 }
 
 .footer .bg-primary {
-  background-color: $primary-450 !important;
+  background-color: $primary-900 !important;
 }
 .footer .card-header:first-child {
   border-radius: 0rem;
@@ -547,10 +519,9 @@ table.td-align-top td {
   margin-bottom: .5rem;
 }
 
-.irp-table th,
-.irp-table td {
+.irp-table th, .irp-table td {
   vertical-align: top;
-  padding: 0.5rem 0.3rem;
+  padding: .5rem .3rem;
 }
 
 .login-logo {
@@ -591,7 +562,6 @@ table.td-align-top td {
   color: $gray-100;
   background-color: $secondary-color;
   border-color: $secondary-color;
-  height: fit-content;
 }
 
 .btn-primary:hover {
@@ -825,7 +795,7 @@ table.td-align-top td {
 }
 
 .btn-level:not(.disabled).active,
-.btn-level:not(.disabled).activeLevel,
+.btn-level:not(.disabled):active,
 .show > .btn-level.dropdown-toggle {
   background-color: $secondary-color;
   color: $white;
@@ -863,7 +833,7 @@ table.td-align-top td {
 }
 
 .ngx-slider .ngx-slider-tick-legend {
-  font-size: 0.7rem;
+  font-size: .7rem;
 }
 
 .ngx-slider .ngx-slider-bubble {
@@ -872,12 +842,12 @@ table.td-align-top td {
 
 .gen-sal-slider-group {
   border: 1px solid $primary-200;
-  border-radius: 0.5rem;
+  border-radius: .5rem;
 }
 
 .sal-slider-label {
   display: inline-block;
-  font-size: 0.6em;
+  font-size: .6em;
   background: $gray-600;
   color:white;
   border-radius: 5px;
@@ -982,7 +952,7 @@ ul.mat-tree-node {
 .debug-highlight {
   background-color: $bg-debug-highlight;
   padding: 3px 4px;
-  border-radius: 0.3rem;
+  border-radius: .3rem;
 }
 
 /* --- spinner styles -- */
@@ -1032,7 +1002,7 @@ div.spinner-size-50 {
 /* spinner fade animation */
 .spinner-fade-in {
   opacity: 0 !important;
-  transition: 0.2s ease-in-out all;
+  transition: .2s ease-in-out all;
 }
 
 .spinner-show {
@@ -1054,7 +1024,7 @@ ul.mat-tree-node mat-tree-node {
 .in-page-message {
   margin: 2em 0;
   border: 1px solid $gray-250;
-  border-radius: 0.5em;
+  border-radius: .5em;
   padding: 1em;
 }
 
@@ -1171,9 +1141,8 @@ th.data-table {
 
 // Header / Navbar
 /////////////////////////////////////////////
-.tsa-header {
+.ncua-header {
   background-color: $primary-color !important;
-  background-color: #215283;
 }
 
 .header-help {
@@ -1250,7 +1219,7 @@ ul.nav-tabs li.active button:hover {
 ul.nav-tabs li button {
   color: $gray-100;
   font-weight: 500;
-  font-size: 1rem;
+  font-size: 1.0rem;
 }
 
 ul.nav-tabs li button:focus,
@@ -1300,7 +1269,8 @@ ul.nav-tabs li.active button .tab-icon {
   padding: 5px;
 }
 .navbar-brand {
-  padding: 0 0 15px 0;
+  padding: 0rem;
+  margin-top: -15px;
 }
 
 .dropdown-menu > li > a {
@@ -1324,6 +1294,7 @@ ul.nav-tabs li.active button .tab-icon {
   background-color: transparent;
   text-decoration: none !important;
 }
+
 
 
 // Menus
@@ -3177,11 +3148,33 @@ hr.ranked-questions {
 //// Maturity Models ////
 
 .maturity-model-card {
-  border-radius: 0.3rem;
+  border: 2px solid $primary-900;
+  background-color: $white;
+  border-radius: .5rem;
+  padding: 1rem;
+  padding-top: .4rem;
+}
+
+input[type=radio].maturity-model-select {
+  margin-right: .5rem;
+}
+
+.maturity-model-select {
+  border: 2px solid $gray-250;
+  border-radius: .3rem;
   margin-bottom: 1rem;
   padding: 1rem;
   cursor: pointer;
 }
+
+.maturity-model-select:hover {
+  background-color: #eff3f6;
+}
+
+.model-selected {
+  border-color: green;
+}
+
 
 /////////////////// Tutorial pages ///////////////
 div.tutorial h5, .tutorial h6 {
@@ -3211,12 +3204,6 @@ div.tutorial h6  {
   margin-top: .2rem;
   margin-bottom: .2rem;
 }
-.max-1000{
-  max-width:1000px;
-}
-.max-1200{
-  max-width:1200px;
-}
 .docTitle{
   font-size: large;
   color: $primary-450;
@@ -3243,7 +3230,7 @@ div.tutorial h6  {
   .example-tree-invisible {
     // display: none;
   }
-
+  
   .example-tree ul,
   .example-tree li {
     margin-top: 0;
@@ -3254,6 +3241,7 @@ div.tutorial h6  {
     font-size: 1rem;
     color: $gray-250;
   }
+
